Refactor GUI initialization and add flag icons
This commit refactors the GUI initialization code in the `CurrencyConverterApp` class to improve readability and maintainability. It also introduces the functionality to display flag icons corresponding to selected currencies in the currency combo boxes.

The `main` method now sets the frame size to 350x400 pixels, providing a better user interface experience. The `convertButton` variable is now declared locally within the `CurrencyConverterApp` constructor, as it is only used within that scope.

The code now includes additional methods: `getFlagIcon`, `getDefaultFlagIcon`, and `getCountryCodeFromCurrency`. These methods are responsible for fetching flag icons from an online source based on the selected currency codes. In case of any errors or if the flag icon is not found, a default flag icon is displayed.

The `getCountryCodeFromCurrency` method retrieves the corresponding country code based on the currency code using the `Locale` class. If a country code is not found, it returns `null`.

Furthermore, the commit modifies the `OpenExchangeRatesConnector` class. The `getExchangeRateData` method is removed as it is no longer used. The `readAPIKey` method is made private and static to improve encapsulation and to be accessible within the class. The unnecessary print statements in the `getExchangeRate` and `getCurrencies` methods are removed.